Update to latest non-migrated package:logging
This is the latest revision that has not been migrated to null safety.
It improves the docs and is worth syncing even though it is not
published.

We cannot bring the null safety migrated version into the SDK yet
because it is not in the allow list for the experiment.

Change-Id: Ie319d81fb4868ed284613424185045f3496e04ad
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/163405
Auto-Submit: Nate Bosch <nbosch@google.com>
Commit-Queue: Nate Bosch <nbosch@google.com>
Commit-Queue: Jake Macdonald <jakemac@google.com>
Reviewed-by: Jake Macdonald <jakemac@google.com>