Initial support for referencing imported names in data-driven fixes
This doesn't yet update the imports when appropriate. To do that I think
I'm going to refactor the code so that code templates and value
generators have a `validate` and `apply` pair (though I might call
`apply` something like `generate`). Right now the code is generating
values during validation (to ensure that values _can_ be generated),
but we can't update the imports during validation.

Change-Id: I97b8cc4e9dfcea6611652f5a78ff664d3b93f761
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/163683
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>