[ dart:io ] Fix socket_connect_consume_write_close_test
The socket connection was being closed prematurely as the SocketServer
was not holding onto a reference to the connected socket. This was
previously fine since the network profiling implementation held on to a
reference within dart:io.

Fixes https://github.com/dart-lang/sdk/issues/43515

Change-Id: If85e0d63beb8299a9fbe86986c3d33a9c1bbb222
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/163901
Reviewed-by: Alexander Aprelev <aam@google.com>
Commit-Queue: Ben Konyi <bkonyi@google.com>