Offer diagnostics for new metadata file to LSP clients
We are in the process of adding a new YAML file for configuring the
fixes in the server, and server is producing diagnostics when the file
content isn't valid. This is an attempt to make those diagnostics
available for clients.

I wasn't sure where to add tests for this, so a pointer would be
appreciated.

Change-Id: Ibd60d70867c6f10835b6f9014006d50b91259549
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/164460
Reviewed-by: Danny Tuppeny <danny@tuppeny.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>