[dart2js] Register `findType` static uses for constants.
We need to register `findType` with the codegen registry whenever we
construct a `TypeReference` so that the function is available at
runtime. We primarily do this whenever we encounter an `HLoadType`.
However, it's possible that we only construct a `TypeReference` during
constant emission, and therefore never register a static use.

We can use the `CodegenImpactTransformer` to register a static use for
`findType` whenever we encounter a `ConstantUse` whose value requires
it. These correspond to the uses of `_reifiedTypeNewRti` in the constant
emitter.

Change-Id: I93dade3c326f7e2c9d4806d8bdc2ee576a20d1c1
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/165480
Commit-Queue: Mayank Patke <fishythefish@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>