[ddc] Fix types of regex match results from js
The value returned from a calling `exec()` is an array-like object
where the values at valid indices can be a string or undefined
(maybe null on some browsers). Changing the element type of the
list to `String?` better reflects that.

As far as I can tell this isn't a problem now because we still treat
local variables initialized with a method call as nullable so the
value is always null checked.

I'm working on a change to start trusting the return type when running
with --sound-null-saftey that exposed the errors in the explicit types.

Change-Id: Ia98fd64549f572a51f0060da3508f4f69557e56c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/165440
Reviewed-by: Mark Zhou <markzipan@google.com>
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>