[tools] Make generate_idefiles ignore most Dart files
Make tools/generate_idefiles.py generate analysis_options.yaml which
ignores most Dart files while we are in a state in which we pass a flag
to the analyzer on whether it should analyze in NNBD mode or not.

Both passing the flag and not passing the flag currently results in
10000s errors in the SDK on the files which should be analyzed the other
way.

This CL changes generate_idefiles.py to exclude many folders, after
which the user can manually uncomment some of the lines in
analysis_options.yaml that correspond the current dartanalyzer flag
setting.

Change-Id: I6c05251691ee1cae12ab3a98fbaf1e6a97b37475
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/168344
Commit-Queue: Daco Harkes <dacoharkes@google.com>
Reviewed-by: Tess Strickland <sstrickl@google.com>