Update Dart.g to match language specification
This CL updates the spec_parser grammar Dart.g such that it corresponds
more precisely to the grammar in the language specification.

Main corrections:
  fieldFormalParameter now allows `?` on the parameter type.
  cascade was reorganized in the spec, Dart.g now reorganized similarly.
  `?` removed from argumentPart (we did not add null-aware invocation)

Other changes were mainly renamings and simplifications of rules, to
make them more similar to the version in the specification.

Change-Id: I1a9bc0097fe6b260dfa6657c947840f2deb437b3
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/167805
Reviewed-by: Lasse R.H. Nielsen <lrn@google.com>
Commit-Queue: Erik Ernst <eernst@google.com>