[vm] Print symbol names and offsets in simulator backtraces.
When running the simulator on precompiled mode, source information
is lost, so only the filename is valid with the line and column
information printed as -1.  For each frame, this CL adds the symbol name
of the instructions section and a hexadecimal offset from the start of
the section, which can be used with the native_stack_traces package. The
absolute address of the VM and isolate instructions are also included
at the start of the backtrace.

Cq-Include-Trybots: luci.dart.try:vm-kernel-precomp-linux-release-simarm-try,vm-kernel-precomp-linux-release-simarm64-try,vm-kernel-linux-release-simarm-try,vm-kernel-linux-release-simarm64-try
Change-Id: Ie67f914b156ead48fccc3069ced6a26940b80673
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/168822
Commit-Queue: Tess Strickland <sstrickl@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>