[CFE] Reproduction of member-signature vs forwarding stub for flutter issue #66122
This reproduces the weirdness seen in
https://github.com/flutter/flutter/issues/66122#issuecomment-714799890

In flutter it furthermore makes the compiled javascript go into a stack
overflow.

Notice how I've also had to introduce the option of disabling the
class ClassHierarchyTest (as it fails) to get both expect files so we
can see the differences, e.g. the introduction of

```
  forwarding-stub set _widget(self2::HotReloadIssue* _) → void
    return super.{fra::State::_widget} = _;
```

in `_HotReloadIssueState extends self2::__HotReloadIssueState&State&AfterLayoutMixin`.

Note that this CL doesn't fix anything.

Change-Id: Ieed6c4a3f36b94ddacc388fb2683b30e88f7d2ba
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/168945
Commit-Queue: Jens Johansen <jensj@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>